==> nts_fmri_project/script02_tms_response_amygdala.ipynb <==
# extract tms response in the amygdala:
# using images generated in last step:

import pandas as pd

## NTS group: use file created in last step:

input_file = "out01_img_file_info_life_time_mdd.csv"
output_file = "out02_img_file_info_mdd_tms_response.csv"

# input_file = "out01_img_file_info_current_mdd.csv"
# output_file = "out02_img_file_info_current_mdd_tms_response.csv"

# input_file = "out01_img_file_info_phq.csv"
# output_file = "out02_img_file_info_phq_tms_response.csv"

data = pd.read_csv(input_file)

data.head()


# list roi mask:
import glob
from nilearn import maskers
import numpy as np

# FSL_R_Amy_thr40
# ROI_R_Amy_Com_MNI
# FIRST_R_amyg_small
# ROI_B_Amy_Com_MNI
# FSL_L_Amy_thr40
# FIRST_B_amyg_small
# FSL_B_Amy_thr40
# ROI_L_Amy_Com_MNI
# FIRST_L_amyg_small

# base_dir = "/Users/jingjiang/Library/CloudStorage/GoogleDrive-jingjiangpsy@gmail.com/Shared drives/Jing_Drive/"
# roi_dir = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/ROI_masks/Amy/"

base_dir = "/home/niki/Downloads/JiangLab/"
roi_dir = "target_mask/"


mask_dir = glob.glob(base_dir + roi_dir + "*.nii")
mask_name = [x.split('/')[-1][:-4] for x in mask_dir]

for m, m_name in zip(mask_dir, mask_name):
    print(m_name)
    tms_site_response = []
    
    for i in range(data.shape[0]):
        image = data.iloc[i, 0]
        nifti_masker = maskers.NiftiMasker(m).fit()
        tms_site_response.append(np.mean(nifti_masker.transform(image)))
        
    data[m_name] = tms_site_response

mask_dir

data.to_csv(output_file, index = None)

## HC group:
# extracing signals for NTHC, TEHC:

import glob

# base_dir = "/Users/jingjiang/Library/CloudStorage/GoogleDrive-jingjiangpsy@gmail.com/Shared drives/Jing_Drive/"
# # img_dir = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/L_pMFG/NTHC_2ndlevel/"

# img_dir1 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/L_pMFG/NTHC_2ndlevel/"
# img_dir2 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/L_aMFG/NTHC_2ndlevel/"
# img_dir3 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/R_M1/NTHC_2ndlevel/"
# img_dir4 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/R_pMFG/NTHC_2ndlevel/"
# img_dir5 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/R_aMFG/NTHC_2ndlevel/"

# img_file_nts = glob.glob(base_dir + img_dir1 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir2 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir3 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir4 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir5 + "CausCon_NTHC*.nii")


base_dir = "/home/niki/Downloads/JiangLab/"
img_dir1 = "tms_response/L_pMFG/??HC_2ndlevel/"
img_dir2 = "tms_response/L_aMFG/??HC_2ndlevel/"
img_dir3 = "tms_response/R_M1/??HC_2ndlevel/"
img_dir4 = "tms_response/R_pMFG/??HC_2ndlevel/"
img_dir5 = "tms_response/R_aMFG/??HC_2ndlevel/"

img_file_hc = glob.glob(base_dir + img_dir1 + "CausCon_??HC*.nii") + \
               glob.glob(base_dir + img_dir2 + "CausCon_??HC*.nii") + \
               glob.glob(base_dir + img_dir3 + "CausCon_??HC*.nii") + \
               glob.glob(base_dir + img_dir4 + "CausCon_??HC*.nii") + \
               glob.glob(base_dir + img_dir5 + "CausCon_??HC*.nii")

file_name = [x.split('/')[-1] for x in img_file_hc]
subject_id = [int(x.split('_')[1][-4:]) for x in file_name]
site = [x.split('_')[3] + '_' + x.split('_')[4] for x in file_name]
subject_id

df_img_file = pd.DataFrame({'file_dir': img_file_hc, 
                            'file_name': file_name,
                            'site': site,
                            'subject_id': subject_id})

print(df_img_file.shape)
df_img_file.head()

# base_dir = "/Users/jingjiang/Library/CloudStorage/GoogleDrive-jingjiangpsy@gmail.com/Shared drives/Jing_Drive/"
# roi_dir = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/ROI_masks/Amy/"


# use new direcotry for mask for dell laptop:
base_dir = "/home/niki/Downloads/JiangLab/"
roi_dir = "target_mask/"

mask_dir = glob.glob(base_dir + roi_dir + "*.nii")
mask_name = [x.split('/')[-1][:-4] for x in mask_dir]

for m, m_name in zip(mask_dir, mask_name):
    print(m_name)
    tms_site_response = []
    
    for i in range(df_img_file.shape[0]):
        image = df_img_file.iloc[i, 0]
        nifti_masker = maskers.NiftiMasker(m).fit()
        tms_site_response.append(np.mean(nifti_masker.transform(image)))
        
    df_img_file[m_name] = tms_site_response


output_file = "out02_img_file_info_hc_tms_response.csv"
df_img_file.to_csv(output_file, index = None)


## HC group:
# extracing signals for NTHC, TEHC:

import glob

# base_dir = "/Users/jingjiang/Library/CloudStorage/GoogleDrive-jingjiangpsy@gmail.com/Shared drives/Jing_Drive/"
# # img_dir = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/L_pMFG/NTHC_2ndlevel/"

# img_dir1 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/L_pMFG/NTHC_2ndlevel/"
# img_dir2 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/L_aMFG/NTHC_2ndlevel/"
# img_dir3 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/R_M1/NTHC_2ndlevel/"
# img_dir4 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/R_pMFG/NTHC_2ndlevel/"
# img_dir5 = "3_Data_Working/CausalConnectome/post_processing/tmsfmri/group_analysis_HRF_all/R_aMFG/NTHC_2ndlevel/"

# img_file_nts = glob.glob(base_dir + img_dir1 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir2 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir3 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir4 + "CausCon_NTHC*.nii") + \
#                glob.glob(base_dir + img_dir5 + "CausCon_NTHC*.nii")

base_dir = "/home/niki/Downloads/JiangLab/"
img_dir1 = "tms_response/L_pMFG/NTS_2ndlevel/"
img_dir2 = "tms_response/L_aMFG/NTS_2ndlevel/"
img_dir3 = "tms_response/R_M1/NTS_2ndlevel/"
img_dir4 = "tms_response/R_pMFG/NTS_2ndlevel/"
img_dir5 = "tms_response/R_aMFG/NTS_2ndlevel/"

img_file_hc = glob.glob(base_dir + img_dir1 + "CausCon_NTS*.nii") + \
               glob.glob(base_dir + img_dir2 + "CausCon_NTS*.nii") + \
               glob.glob(base_dir + img_dir3 + "CausCon_NTS*.nii") + \
               glob.glob(base_dir + img_dir4 + "CausCon_NTS*.nii") + \
               glob.glob(base_dir + img_dir5 + "CausCon_NTS*.nii")

file_name = [x.split('/')[-1] for x in img_file_hc]
subject_id = [int(x.split('_')[1][-4:]) for x in file_name]
site = [x.split('_')[3] + '_' + x.split('_')[4] for x in file_name]
subject_id

df_img_file = pd.DataFrame({'file_dir': img_file_hc, 
                            'file_name': file_name,
                            'site': site,
                            'subject_id': subject_id})

print(df_img_file.shape)
df_img_file.head()

# use new direcotry for mask for dell laptop:
base_dir = "/home/niki/Downloads/JiangLab/"
roi_dir = "target_mask/"

mask_dir = glob.glob(base_dir + roi_dir + "*.nii")
mask_name = [x.split('/')[-1][:-4] for x in mask_dir]

for m, m_name in zip(mask_dir, mask_name):
    print(m_name)
    tms_site_response = []
    
    for i in range(df_img_file.shape[0]):
        image = df_img_file.iloc[i, 0]
        nifti_masker = maskers.NiftiMasker(m).fit()
        tms_site_response.append(np.mean(nifti_masker.transform(image)))
        
    df_img_file[m_name] = tms_site_response


output_file = "out02_img_file_info_nts_tms_response.csv"
df_img_file.to_csv(output_file, index = None)
