==> nts_fmri_project/script04_stats_mdd_nthc.Rmd <==
---
title: "R Notebook"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
---

#```{r}
#library("MatchIt")
#data = read.csv('out03_data_ltmdd_nthc.csv', sep = ',')
#
#data$group2 = 1
#data[data$group == "nthc", 'group2'] <- 0
#
#dim(data)
#
#m.out1 <- matchit(group2 ~ age + gender + yrs_of_edu , data = data,
#                 method = "nearest", distance = "glm",
#                 #method = "full", distance = "glm", link = "probit",
#                 )
#
#plot(m.out1, type = "jitter", interactive = FALSE)
#
#m.data <- match.data(m.out1)
#m.data = subset(m.data, select = -c(group2))
#write.csv(x = m.data, file = "out04_data_ltmdd_nthc_match.csv", row.names = FALSE)
#```


```{r}
library(lme4)
library(reshape2)
library(lmerTest)

roi_right = c("FSL_R_Amy_thr40", 
            "ROI_R_Amy_Com_MNI",
            "FIRST_R_amyg_small")

roi_left = c("FSL_L_Amy_thr40",
            "ROI_L_Amy_Com_MNI",
            "FIRST_L_amyg_small")

data = read.csv('out03_data_ltmdd_nthc.csv', sep = ',')

#data.lpmfg = data[data$site == 'L_pMFG', ]
#response ～ roi + group + sites + (1 | subID)


data = melt(data, id=c("subject_id","group", "site", "age", "gender", "yrs_of_edu"), value.name = 'response', variable.name = 'roi')

data = data[data$site != 'R_M1', ]

data$subject_id = as.factor(data$subject_id)
data$site = as.factor(data$site)
data$group = as.factor(data$group)
data$group = as.factor(data$group)


for (i in 1:length(roi_left)) {
  data.i = rbind(data[data$roi == roi_right[i], ],
                 data[data$roi == roi_left[i], ])
  
  model <- lmer(response ~ group * roi * site + (1 | subject_id), data=data.i, REML = FALSE,
                control = lmerControl(optimizer = "bobyqa"))
  print(summary(model))
  print(anova(model))
}


```
Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Cmd+Option+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Cmd+Shift+K* to preview the HTML file). 

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.
